# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
      "AIAnalysisRequest": {
        "properties": {
          "agent_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Code"
          },
          "focus": {
            "default": "global",
            "title": "Focus",
            "type": "string"
          },
          "simulation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "window_days": {
            "default": 30,
            "title": "Window Days",
            "type": "integer"
          }
        },
        "title": "AIAnalysisRequest",
        "type": "object"
      },
...
      "CausalAnalysisRequest": {
        "properties": {
          "event_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type"
          },
          "horizon_days": {
            "default": 20,
            "title": "Horizon Days",
            "type": "integer"
          },
          "region": {
            "default": "global",
            "title": "Region",
            "type": "string"
          },
          "use_ai": {
            "default": true,
            "title": "Use Ai",
            "type": "boolean"
          },
          "window_days": {
            "default": 30,
            "title": "Window Days",
            "type": "integer"
          }
        },
        "title": "CausalAnalysisRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportExport": {
        "properties": {
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "markdown"
        ],
        "title": "ReportExport",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "全球多源综合风险指数与世界局势情景模拟器",
    "title": "WorldPulse",
    "version": "1.2.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ai/analyze-risk": {
      "post": {
        "operationId": "ai_analyze_risk_api_ai_analyze_risk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Analyze Risk"
      }
    },
    "/api/ai/analyze-simulation": {
      "post": {
        "operationId": "ai_analyze_simulation_api_ai_analyze_simulation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Analyze Simulation"
      }
    },
    "/api/ai/report/export": {
      "post": {
        "operationId": "ai_report_export_api_ai_report_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ReportExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Report Export"
      }
    },
...
    "/api/causal/analyze": {
      "post": {
        "operationId": "causal_analyze_api_causal_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CausalAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CausalAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Causal Analyze"
      }
    },
    "/api/causal/backtest": {
      "get": {
        "operationId": "causal_backtest_api_causal_backtest_get",
        "parameters": [
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "default": "conflict",
              "title": "Event Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 120,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "horizon_days",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Horizon Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CausalBacktestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Causal Backtest"
      }
    },
    "/api/causal/events": {
      "get": {
        "operationId": "causal_events_api_causal_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "global",
              "title": "Region",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CausalEvent"
                  },
                  "title": "Response Causal Events Api Causal Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Causal Events"
      }
    },
    "/api/causal/graph": {
      "get": {
        "operationId": "causal_graph_api_causal_graph_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "global",
              "title": "Region",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/CausalChain"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Causal Graph"
      }
    },
    "/api/causal/report/export": {
      "post": {
        "operationId": "causal_report_export_api_causal_report_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CausalAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ReportExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Causal Report Export"
      }
    },
    "/api/events/agent/{code}": {
      "get": {
        "operationId": "events_agent_digest_api_events_agent__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Window Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/EventDigest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events Agent Digest"
      }
    },
    "/api/events/digest": {
      "get": {
        "operationId": "events_digest_api_events_digest_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "global",
              "title": "Region",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/EventDigest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events Digest"
      }
    },
...
    "/api/exports/replay.csv": {
      "get": {
        "operationId": "export_replay_csv_api_exports_replay_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 120,
              "title": "Window Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Replay Csv"
      }
    },
...
    "/api/exports/species-occurrences.csv": {
      "get": {
        "operationId": "export_species_occurrences_csv_api_exports_species_occurrences_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "scientific_name",
            "required": true,
            "schema": {
              "title": "Scientific Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "gbif",
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "africa",
              "title": "Region",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 240,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json; charset=utf-8": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Species Occurrences Csv"
      }
    },
...
  }
}